Flatten app package
